// src/device/enums/pressure-device-slug.enum.ts
export enum PressureDeviceSlug {
  G05_GOODS_OUT_PRESSURE = 'g05-goods-out-pressure',
  G07_QUARANTINE_STORE_PRESSURE = 'g07-quarantine-store-pressure',
  G08_RELEASED_MATERIALS_STORE_PRESSURE = 'g08-released-materials-store-pressure',
  G23_SPRAY_DRYING_PRESSURE = 'g23-spray-drying-pressure',
  G26_POWDER_BOOTH_PRESSURE = 'g26-powder-booth-pressure',
  G28_NON_HP_CLEANROOM_PRESSURE = 'g28-non-hp-cleanroom-pressure',
  G41_HP3_HIGH_POTENCY_PRESSURE = 'g41-hp3-high-potency-pressure',
  G37_HP2_HIGH_POTENCY_PRESSURE = 'g37-hp2-high-potency-pressure',
  G33_HP1_HIGH_POTENCY_PRESSURE = 'g33-hp1-high-potency-pressure',
}